added a test user with hashed password to seeder, updated the functions to work in profiles, users.create changed to work with models, user.edit was just testing tbh, updated Auth to only hold functions that it needs
